fix: 파일명, 코드를 컴포넌트로 변경
